requirements.txt
# Core requirements
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.39.0  # For quantization support
peft>=0.4.0  # For parameter-efficient fine-tuning

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image processing
pillow>=9.5.0
opencv-python>=4.7.0

# Notebook support
jupyterlab>=4.0.0
notebook>=6.5.0

# Utilities
tqdm>=4.65.0  # Progress bars
pytest>=7.3.0  # Testing

# Optional: for more advanced models
# sentencepiece>=0.1.99  # For tokenization
# protobuf>=3.20.0  # Required for some model loaders
# safetensors>=0.3.1  # For safer model loading